govly-web/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.0
requests>=2.32.5
pydantic>=2.11.5,<3
supabase==2.0.2
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.1
sentence-transformers>=3.1.1,<3.5
torch>=2.2.0,<3
transformers>=4.41,<5
numpy<2.0.0

# LangChain - use compatible versions
langchain>=0.3.27
langchain-core>=0.3.75
langchain-community>=0.3.29 

# LlamaIndex (aligned versions)
llama-index==0.13.5
# extensions compatible with 0.13.x core
llama-index-embeddings-huggingface>=0.6.0,<0.7
llama-index-vector-stores-supabase>=0.4.0,<0.5

# AWS SDK for Textract
boto3>=1.34.0,<2.0
botocore>=1.34.0,<2.0